extract options processing
